Agumentik Group of Companies
Agumentik Group of Companies
Task Details:
Here is the Figma URL for your design reference where you can use this for making the responsive frontend design here in the design there are several pages where compulsory to create 3 pages which 1 page should be the home page and the other 2 pages you have to choose as per your choice from this design and make it ready.
 
Parameters of the task: (Note: Read the parameters carefully, you will be judged on the basis of parameters only)
For Frontend:
The design should be responsive, as it is in Figma.
The code structure should be clean.
Code should be well documented (proper comments wherever required for code module ideas, etc.).
 
For Backend:
All images and content should come from the Admin Panel, which will be dynamic and allow us to change them at any time.
Create an option to display social icons such as Facebook, LinkedIn, and Instagram at two locations of the page/Screen, and both should be updated from the admin panel from one location (if we change the Facebook link from the admin panel in both locations, the link should be changed at the same time).
Create a contact form. It should be eye-catching and once someone clicks, one popup should come out, and, with that name and contact number, people will submit, which will be saved in the admin panel as leads, which the admin can see or download in pdf. There should be a thank you email to the visitors who will submit information in the popup. So, add an admin option for logging in as an admin. In this way, an existing admin can add another admin.